requirements.txt
datasets
loguru
umap-learn
sentence-transformers